backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
cachetools==5.3.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0.6
dnspython==2.6.1
docker-pycreds==0.4.0
email_validator==2.1.1
eval_type_backport==0.2.0
executing==2.0.1
faiss-cpu==1.8.0.post1
fastjsonschema==2.18.0
feedparser==6.0.11
filelock==3.15.3
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.4
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.0
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.23
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93
markdown-it-py==3.0.0
markdownify==0.13.1
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.1
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.25.1
oauthlib==3.2.2
openai==1.37.0
orjson==3.10.6
overrides==7.4.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
prometheus-client==0.17.1
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.18.0
PyJWT==2.8.0
pypandoc==1.13
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.30.2
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
semanticscholar==0.8.4
Send2Trash==1.8.2
sentry-sdk==1.45.0
setproctitle==1.3.3
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.5.0
terminado==0.17.1
threadpoolctl==3.4.0
tiktoken==0.7.0
tinycss2==1.2.1
together==1.2.1
tokenizers==0.19.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.41.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
wandb==0.16.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4
